requirements.txt
cloudinary==1.42.1
Flask==2.2.2
python-dotenv==0.21.0
deepeval==2.0.9
diffusers==0.30.3
docx2txt==0.8
execnet==2.1.1
googleapis-common-protos==1.66.0
jsonpointer==2.1
marshmallow==3.23.2
migrate==0.3.8
nltk==3.9.1
protobuf==3.20.3
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyls-spyder==0.4.0
pypdf==5.1.0
PyPDF2==3.0.1
PyQt5==5.15.10
PyQtWebEngine==5.15.6